Create 424-Longest-Repeating-Character-Replacement.py
Time Complex. = O(26n)
Space Complex. = O(26)
Possible to get time complexity of O(n) because we only need to get the max frequency whenever we update the array by adding to a num. We only get the max length when max frequency occurs, therefore we do not care to update the max frequency when we delete from the array.
7/20/22