Create 108-Convert-Sorted-Array-to-Binary-Search-Tree
Time Complex. = O(n) since each non-null node will be cycled through once
Space Complex. = O(lgn)
The Reason why space complexity is O(lgn) is because we are doing DFS with a stack, so at all times we will have at most the height of the tree in our stack multiplied by the branching factor, which is 2. So since the height of a BST is lgn and the branching factor is 2, we get O(2lgn) or O(lgn). Technically there is a way to get O(lgn) even with a branching factor of 2, but it sacrifices time complexity as we will have to expand every node twice.
7/22/22